Take reference when re-using already opened display
This way, the display should get closed after the last call to close().

Note: creating new GObject will initialize the refcount to 1, so no *additional* g_object_ref is wanted in this case. As per Gnome documentation:

"The reference count is initialized to one by g_object_new which means that the caller is currently the sole owner of the newly-created reference. When the reference count reaches zero, that is, when g_object_unref is called by the last client holding a reference to the object, the dispose and the finalize class methods are invoked."

(Finalize is what will close the native display handle for us.)