Check if the compile-time version of GStreamer matches the run-time one
The GL code currently isn't stable, and indeed broke ABI compatibility between 1.8 and 1.10. That's normally not a problem for systems were we're shipping a build of GStreamer, except on (x86) Linux, were we're currently relying on the host having a copy installed. For this case especially, print a warning if the versions don't match.